# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Sample API",
        "contact": {
            "name": "Erik Alexander, Friscilia Sutanto, Sri Wardani",
            "url": "https://uph.edu"
        },
        "version": "1.0"
    },
    "host": "webservice.app",
    "basePath": "",
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/api/v1/nilai": {
            "get": {
                "tags": [
                    "Nilai"
                ],
                "summary": "Retrieves the collection of Nilai resources.",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Nilai collection.",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/nilai"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized action."
                    }
                }
            }
        },
        "/api/nilai": {
            "post": {
                "tags": [
                    "Nilai"
                ],
                "summary": "Data Nilai",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "Data Jurusan",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Users Token.",
                        "schema": {
                            "property": "token",
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized action."
                    }
                }
            }
        },
        "/api/v1/nilai/{id}": {
            "get": {
                "tags": [
                    "Nilai"
                ],
                "summary": "Retrieves the collection of Nilai resources.",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Nilai collection.",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/nilai"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized action."
                    }
                }
            },
            "put": {
                "tags": [
                    "Nilai"
                ],
                "summary": "Data Nilai",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "Data Nilai",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success Update.",
                        "schema": {
                            "property": "token",
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized action."
                    }
                }
            },
            "delete": {
                "tags": [
                    "Nilai"
                ],
                "summary": "Removes the Nilai resource.",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Nilai resource delete."
                    },
                    "401": {
                        "description": "Unauthorized Action."
                    },
                    "404": {
                        "description": "Resource not found."
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "nilai": {
            "properties": {
                "semester": {
                    "type": "string"
                },
                "id_dosen": {
                    "type": "string"
                },
                "id_matkul": {
                    "type": "integer"
                },
                "id_mahasiswa": {
                    "type": "integer"
                },
                "absensi": {
                    "type": "string"
                },
                "tugas": {
                    "type": "string"
                },
                "uts": {
                    "type": "string"
                },
                "uas": {
                    "type": "string"
                },
                "grade": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "update_at": {
                    "type": "string"
                }
            }
        }
    }
}